+++ CSharp_Engine/CSharp_Engine.csproj
@@ -5,11 +5,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{E35FD3CF-5B58-49C2-B44A-E97F656F009B}</ProjectGuid>
+    <ProjectGuid>{3FA232F2-1078-4682-B402-CE20B9665856}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>BH.oM.Node2Code</RootNamespace>
-    <AssemblyName>Node2Code_oM</AssemblyName>
+    <RootNamespace>BH.Engine.CSharp</RootNamespace>
+    <AssemblyName>CSharp_Engine</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -39,14 +39,30 @@
       <HintPath>..\..\BHoM\Build\BHoM.dll</HintPath>
       <Private>False</Private>
     </Reference>
+    <Reference Include="Library_Engine">
+      <HintPath>..\..\BHoM_Engine\Build\Library_Engine.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
     <Reference Include="Microsoft.CodeAnalysis, Version=3.4.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
-      <HintPath>..\packages\Microsoft.CodeAnalysis.Common.3.4.0-beta4-final\lib\netstandard2.0\Microsoft.CodeAnalysis.dll</HintPath>
+      <HintPath>..\packages\Microsoft.CodeAnalysis.Common.3.4.0-beta3-final\lib\netstandard2.0\Microsoft.CodeAnalysis.dll</HintPath>
     </Reference>
     <Reference Include="Microsoft.CodeAnalysis.CSharp, Version=3.4.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
-      <HintPath>..\packages\Microsoft.CodeAnalysis.CSharp.3.4.0-beta4-final\lib\netstandard2.0\Microsoft.CodeAnalysis.CSharp.dll</HintPath>
+      <HintPath>..\packages\Microsoft.CodeAnalysis.CSharp.3.4.0-beta3-final\lib\netstandard2.0\Microsoft.CodeAnalysis.CSharp.dll</HintPath>
+    </Reference>
+    <Reference Include="Programming_Engine">
+      <HintPath>..\..\BHoM_Engine\Build\Programming_Engine.dll</HintPath>
+      <Private>False</Private>
     </Reference>
     <Reference Include="Programming_oM">
       <HintPath>..\..\BHoM\Build\Programming_oM.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="Reflection_Engine">
+      <HintPath>..\..\BHoM_Engine\Build\Reflection_Engine.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="Reflection_oM">
+      <HintPath>..\..\BHoM\Build\Reflection_oM.dll</HintPath>
       <Private>False</Private>
     </Reference>
     <Reference Include="System" />
@@ -84,8 +100,30 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Others\Variable.cs" />
+    <Compile Include="Compute\Variables.cs" />
+    <Compile Include="Compute\NodeSequence.cs" />
+    <Compile Include="Compute\ApplyGroups.cs" />
+    <Compile Include="Convert\ToCSharpText.cs" />
+    <Compile Include="Convert\ToCSharpSyntaxNode.cs" />
+    <Compile Include="Objects\NodeState.cs" />
+    <Compile Include="Objects\ReceiverState.cs" />
+    <Compile Include="Query\Children.cs" />
+    <Compile Include="Query\Expression.cs" />
+    <Compile Include="Query\SyntaxToken.cs" />
+    <Compile Include="Query\DepthDifference.cs" />
+    <Compile Include="Query\Statements.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Query\OutputVariables.cs" />
+    <Compile Include="Query\Body.cs" />
+    <Compile Include="Query\ArgumentValue.cs" />
+    <Compile Include="Query\ReturnType.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\CSharp_oM\CSharp_oM.csproj">
+      <Project>{e35fd3cf-5b58-49c2-b44a-e97f656f009b}</Project>
+      <Name>CSharp_oM</Name>
+      <Private>False</Private>
+    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -93,6 +131,10 @@
   <ItemGroup>
     <Analyzer Include="..\packages\Microsoft.CodeAnalysis.Analyzers.2.9.6\analyzers\dotnet\cs\Microsoft.CodeAnalysis.Analyzers.dll" />
     <Analyzer Include="..\packages\Microsoft.CodeAnalysis.Analyzers.2.9.6\analyzers\dotnet\cs\Microsoft.CodeAnalysis.CSharp.Analyzers.dll" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Create\" />
+    <Folder Include="Modify\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
