+++ CSharp_Engine/Properties/AssemblyInfo.cs
@@ -27,11 +27,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("Node2Code_oM")]
+[assembly: AssemblyTitle("CSharp_Engine")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("Node2Code_oM")]
+[assembly: AssemblyProduct("CSharp_Engine")]
 [assembly: AssemblyCopyright("Copyright © https://github.com/BHoM")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
@@ -42,7 +42,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("e35fd3cf-5b58-49c2-b44a-e97f656f009b")]
+[assembly: Guid("3fa232f2-1078-4682-b402-ce20b9665856")]
 
 // Version information for an assembly consists of the following four values:
 //
